requirements.txt
pandas
plotly_express
streamlit
scipy
statsmodels
matplotlib
lxml